+++ packages/adapter-kysely/src/utils.ts
@@ -1,8 +1,8 @@
 import type { SessionSchema } from "lucia-auth";
 import type { Selectable } from "kysely";
-import type { Session } from "./dbTypes";
+import type { KyselySession } from "./types.js";
 
-export const convertSession = (session: Selectable<Session>): SessionSchema => {
+export const convertSession = (session: Selectable<KyselySession>): SessionSchema => {
 	const { expires, idle_expires: idleExpires, ...data } = session;
 	return {
 		expires: Number(expires),