+++ packages/adapter-kysely/test/sqlite/index.ts
@@ -1,0 +1,4 @@
+import { testAdapter } from "@lucia-auth/adapter-test";
+import { db, adapter } from "./db.js";
+
+testAdapter(adapter, db);